+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Remove reference to unused module (com.google.gwt.rpc.RPC) that is not present in GWT 2.7.
 
 ## 0.5.28:
+* Support the ability to debug just a single DataSourceLoader services
+  changes.
 * Add ReplicantDev.gwt.xml configuration that turns on all debug features.
 * Exclude test classes form GWT compiler's path.