+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.30 (Pending):
-* Add EntitySubscriptionDebugger to support debugging the set of subscriptions.
+* Add support for debugging entity subscriptions and requests as well as local
+  repository state.
 * If repository debug output is enabled in application then print out a helper message
   when the GwtDataLoaderService is defined. Helps developers remember this feature is
   present.